# src/main/webapp/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Descriptions de toutes les APIs exposées par le bouchon",
    "version" : "v1",
    "title" : "Spécification auto-généré des APIs exposées",
    "termsOfService" : "FMI Corp"
  },
  "host" : "localhost:8080",
  "basePath" : "/v1/coupons",
  "tags" : [ {
    "name" : "coupons"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/v1/coupons" : {
      "get" : {
        "tags" : [ "coupons" ],
        "summary" : "Récupère les coupons",
        "description" : "Permet de récupérer tous les coupons non utilisés",
        "operationId" : "getAllCoupons",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK - La liste de coupons est retournée au format JSON dans le response body",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Coupon"
              }
            }
          },
          "500" : {
            "description" : "Si une erreur se produit"
          }
        }
      },
      "post" : {
        "tags" : [ "coupons" ],
        "summary" : "Ajouter un coupon",
        "description" : "Permet d'ajouter un coupon",
        "operationId" : "addCoupon",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Coupon à ajouter",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Coupon"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Coupon"
            }
          },
          "201" : {
            "description" : "CREATE - Le coupon créé est retourné",
            "schema" : {
              "$ref" : "#/definitions/Coupon"
            }
          },
          "500" : {
            "description" : "Si une erreur se produit"
          }
        }
      }
    },
    "/v1/coupons/{couponId}" : {
      "get" : {
        "tags" : [ "coupons" ],
        "summary" : "Récupère un coupon",
        "description" : "Permet de récupérer un coupon par son identifiant",
        "operationId" : "getCoupon",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "couponId",
          "in" : "path",
          "description" : "Identifiant du coupon à récupérer",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK - Le coupon est retourné au format JSON dans le response body",
            "schema" : {
              "$ref" : "#/definitions/Coupon"
            }
          },
          "404" : {
            "description" : "Coupon non trouvé"
          },
          "500" : {
            "description" : "Si une erreur se produit"
          }
        }
      },
      "delete" : {
        "tags" : [ "coupons" ],
        "summary" : "Supprimer un coupon",
        "description" : "Permet de supprimer un coupon par son identifiant",
        "operationId" : "deleteCoupon",
        "parameters" : [ {
          "name" : "couponId",
          "in" : "path",
          "description" : "Identifiant du coupon à supprimer",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          },
          "204" : {
            "description" : "NO CONTENT"
          },
          "404" : {
            "description" : "Coupon non trouvé"
          },
          "500" : {
            "description" : "Si une erreur se produit"
          }
        }
      }
    }
  },
  "definitions" : {
    "Coupon" : {
      "type" : "object",
      "required" : [ "nom", "reduction" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Identifiant unique du coupon. Null si le coupon n'a pas été posté"
        },
        "nom" : {
          "type" : "string",
          "position" : 1,
          "description" : "Nom du coupon"
        },
        "reduction" : {
          "type" : "string",
          "position" : 2,
          "description" : "Réduction du coupon"
        },
        "estUtilise" : {
          "type" : "boolean",
          "position" : 3,
          "description" : "True si le coupon a été utilisé et false sinon",
          "default" : false
        }
      },
      "description" : "Model de coupon échangé pendant les appels"
    }
  }
}